+++ .gitignore
@@ -286,3 +286,7 @@ __pycache__/
 *.btm.cs
 *.odx.cs
 *.xsd.cs
+/out/build/x64-Debug
+/CMakePresets.json
+/CMakeUserPresets.json
+/lab2/out/build/x64-Debug
