+++ lab2/CMakeLists.txt
@@ -17,6 +17,8 @@ set(PROJECT_SOURCES
         main.cpp
         mainwindow.cpp
         mainwindow.h
+        solver.h
+        solver.cpp
 )
 
 if(${QT_VERSION_MAJOR} GREATER_EQUAL 6)
