+++ lab2/solver.h
@@ -1,0 +1,6 @@
+#include<vector>
+
+
+void run(double ae1, double mu1, double ae2, double mu2, std::vector<double>& a, std::vector<double>& b, std::vector<double>& c, std::vector<double>& phi, std::vector<double>& result);
+
+void buildLES(int n, double& ae1, double& mu1, double& ae2, double& mu2, std::vector<double>& a, std::vector<double>& b, std::vector<double>& c, std::vector<double>& phi);
